requirements.txt
click==8.0.3
pydantic>=2.11.7
pydantic-settings>=2.7.1
pytest==8.3.5
pytest-asyncio>=0.23.0
python-dotenv>=1.1.0
fastmcp==2.14.4
atlassian-python-api>=3.41.0
